unify xp estimator
